+++ src/upc.rs
@@ -25,7 +25,7 @@ impl UPC
     {
         if barcode.len() != 12
         {
-            return Err(UPCParseError::Length(barcode.len() as u8));
+            return Err(UPCParseError::Length(barcode.len()));
         }
 
         let e = UPC { code: barcode.parse::<u64>()?, };
@@ -84,7 +84,7 @@ impl fmt::Display for UPC
 #[derive(Debug)]
 pub enum UPCParseError
 {
-    Length(u8),
+    Length(usize),
     Parse(ParseIntError),
     CheckDigit(u8, u8),
 }
